refactor: move file and update tests
