+++ __mocks__/src/web3/utils.ts
@@ -1,0 +1,4 @@
+module.exports = {
+  ...jest.requireActual('src/web3/utils'),
+  getSupportedNetworkIds: jest.fn().mockReturnValue(['celo-alfajores', 'ethereum-sepolia']),
+}
