+++ src/analytics/AppAnalytics.test.ts
@@ -4,8 +4,9 @@ import AppAnalyticsModule from 'src/analytics/AppAnalytics'
 import { OnboardingEvents } from 'src/analytics/Events'
 import * as config from 'src/config'
 import { store } from 'src/redux/store'
-import { getDefaultStatsigUser, getFeatureGate, getSupportedNetworkIds } from 'src/statsig'
+import { getDefaultStatsigUser, getFeatureGate } from 'src/statsig'
 import { NetworkId } from 'src/transactions/types'
+import { getSupportedNetworkIds } from 'src/web3/utils'
 import { Statsig } from 'statsig-react-native'
 import { getMockStoreData } from 'test/utils'
 import {
@@ -39,6 +40,11 @@ jest.mock('src/web3/networkConfig', () => {
     },
   }
 })
+jest.mock('src/web3/utils', () => ({
+  ...jest.requireActual('src/web3/utils'),
+  getSupportedNetworkIds: jest.fn(),
+}))
+jest.mocked(getSupportedNetworkIds).mockReturnValue([NetworkId['celo-alfajores']])
 
 const mockDeviceId = 'abc-def-123' // mocked in __mocks__/react-native-device-info.ts (but importing from that file causes weird errors)
 const expectedSessionId = '453e535d43b22002185f316d5b41561010d9224580bfb608da132e74b128227a'
@@ -210,7 +216,6 @@ describe('AppAnalytics', () => {
     mockConfig.STATSIG_ENABLED = true
     mockStore.getState.mockImplementation(() => state)
     jest.mocked(getFeatureGate).mockReturnValue(true)
-    jest.mocked(getSupportedNetworkIds).mockReturnValue([NetworkId['celo-alfajores']])
   })
 
   describe('init', () => {
