+++ src/analytics/saga.ts
@@ -1,6 +1,6 @@
 import AppAnalytics from 'src/analytics/AppAnalytics'
 import { getCurrentUserTraits } from 'src/analytics/selectors'
-import { getSupportedNetworkIds } from 'src/statsig'
+import { getSupportedNetworkIds } from 'src/web3/utils'
 import { call, select, spawn, take } from 'typed-redux-saga'
 
 export function* updateUserTraits() {
