+++ src/qrcode/QRCode.tsx
@@ -14,7 +14,6 @@ import StyledQRCode from 'src/qrcode/StyledQRCode'
 import { useSelector } from 'src/redux/hooks'
 import { SVG } from 'src/send/actions'
 import { NETWORK_NAMES } from 'src/shared/conts'
-import { getSupportedNetworkIds } from 'src/statsig'
 import colors from 'src/styles/colors'
 import { typeScale } from 'src/styles/fonts'
 import { vibrateInformative } from 'src/styles/hapticFeedback'
@@ -23,6 +22,7 @@ import variables from 'src/styles/variables'
 import { NetworkId } from 'src/transactions/types'
 import Logger from 'src/utils/Logger'
 import { walletAddressSelector } from 'src/web3/selectors'
+import { getSupportedNetworkIds } from 'src/web3/utils'
 
 interface Props {
   qrSvgRef: React.MutableRefObject<SVG>
