+++ src/send/SendEnterAmount.test.tsx
@@ -10,8 +10,6 @@ import { Screens } from 'src/navigator/Screens'
 import { RecipientType } from 'src/recipients/recipient'
 import SendEnterAmount from 'src/send/SendEnterAmount'
 import { usePrepareSendTransactions } from 'src/send/usePrepareSendTransactions'
-import { getSupportedNetworkIds } from 'src/statsig'
-import { NetworkId } from 'src/transactions/types'
 import { PreparedTransactionsPossible } from 'src/viem/prepareTransactions'
 import MockedNavigator from 'test/MockedNavigator'
 import { createMockStore, mockStoreBalancesToTokenBalances } from 'test/utils'
@@ -94,7 +92,6 @@ const params = {
 describe('SendEnterAmount', () => {
   beforeEach(() => {
     jest.clearAllMocks()
-    jest.mocked(getSupportedNetworkIds).mockReturnValue([NetworkId['celo-alfajores']])
   })
 
   it('should render only the allowed send tokens', () => {
