+++ src/send/utils.ts
@@ -13,12 +13,12 @@ import { UriData, uriDataFromUrl } from 'src/qrcode/schema'
 import { AddressRecipient, Recipient, RecipientType } from 'src/recipients/recipient'
 import { updateAppRecipientCache } from 'src/recipients/reducer'
 import { TransactionDataInput } from 'src/send/types'
-import { getSupportedNetworkIds } from 'src/statsig'
 import { tokensListSelector } from 'src/tokens/selectors'
 import { TokenBalance } from 'src/tokens/slice'
 import { convertLocalToTokenAmount } from 'src/tokens/utils'
 import { Currency } from 'src/utils/currencies'
 import Logger from 'src/utils/Logger'
+import { getSupportedNetworkIds } from 'src/web3/utils'
 import { call, put, select } from 'typed-redux-saga'
 
 const TAG = 'send/utils'
