+++ src/swap/SwapScreen.test.tsx
@@ -12,12 +12,7 @@ import { ErrorMessages } from 'src/app/ErrorMessages'
 import { APPROX_SYMBOL } from 'src/components/TokenEnterAmount'
 import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
-import {
-  getDynamicConfigParams,
-  getExperimentParams,
-  getFeatureGate,
-  getSupportedNetworkIds,
-} from 'src/statsig'
+import { getDynamicConfigParams, getExperimentParams, getFeatureGate } from 'src/statsig'
 import { StatsigFeatureGates } from 'src/statsig/types'
 import SwapScreen from 'src/swap/SwapScreen'
 import { swapStart } from 'src/swap/slice'
@@ -322,9 +317,6 @@ describe('SwapScreen', () => {
     jest.mocked(getExperimentParams).mockReturnValue({
       swapBuyAmountEnabled: true,
     })
-    jest
-      .mocked(getSupportedNetworkIds)
-      .mockReturnValue([NetworkId['celo-alfajores'], NetworkId['ethereum-sepolia']])
     jest.mocked(getDynamicConfigParams).mockReturnValue({
       maxSlippagePercentage: '0.3',
       popularTokenIds: [],
@@ -1806,9 +1798,6 @@ describe('SwapScreen', () => {
 
     it('should show "popular" tokens', () => {
       const mockedPopularTokens = [mockUSDCTokenId, mockPoofTokenId]
-      jest
-        .mocked(getSupportedNetworkIds)
-        .mockReturnValue([NetworkId['celo-alfajores'], NetworkId['ethereum-sepolia']])
       jest.mocked(getDynamicConfigParams).mockReturnValue({
         popularTokenIds: mockedPopularTokens,
         maxSlippagePercentage: '0.3',
