+++ src/swap/SwapScreenV2.test.tsx
@@ -13,12 +13,7 @@ import { APPROX_SYMBOL } from 'src/components/TokenEnterAmount'
 import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
 import { NETWORK_NAMES } from 'src/shared/conts'
-import {
-  getDynamicConfigParams,
-  getExperimentParams,
-  getFeatureGate,
-  getSupportedNetworkIds,
-} from 'src/statsig'
+import { getDynamicConfigParams, getExperimentParams, getFeatureGate } from 'src/statsig'
 import { StatsigFeatureGates } from 'src/statsig/types'
 import SwapScreenV2 from 'src/swap/SwapScreenV2'
 import { swapStart } from 'src/swap/slice'
@@ -327,9 +322,6 @@ describe('SwapScreen', () => {
     jest.mocked(getExperimentParams).mockReturnValue({
       swapBuyAmountEnabled: true,
     })
-    jest
-      .mocked(getSupportedNetworkIds)
-      .mockReturnValue([NetworkId['celo-alfajores'], NetworkId['ethereum-sepolia']])
     jest.mocked(getDynamicConfigParams).mockReturnValue({
       maxSlippagePercentage: '0.3',
       popularTokenIds: [],
@@ -1922,9 +1914,6 @@ describe('SwapScreen', () => {
 
     it('should show "popular" tokens', () => {
       const mockedPopularTokens = [mockUSDCTokenId, mockPoofTokenId]
-      jest
-        .mocked(getSupportedNetworkIds)
-        .mockReturnValue([NetworkId['celo-alfajores'], NetworkId['ethereum-sepolia']])
       jest.mocked(getDynamicConfigParams).mockReturnValue({
         popularTokenIds: mockedPopularTokens,
         maxSlippagePercentage: '0.3',
