+++ src/swap/useFilterChips.ts
@@ -2,7 +2,7 @@ import { useTranslation } from 'react-i18next'
 import { FilterChip } from 'src/components/FilterChipsCarousel'
 import { TOKEN_MIN_AMOUNT } from 'src/config'
 import { useSelector } from 'src/redux/hooks'
-import { getDynamicConfigParams, getFeatureGate, getSupportedNetworkIds } from 'src/statsig'
+import { getDynamicConfigParams, getFeatureGate } from 'src/statsig'
 import { DynamicConfigs } from 'src/statsig/constants'
 import { StatsigDynamicConfigs, StatsigFeatureGates } from 'src/statsig/types'
 import { lastSwappedSelector } from 'src/swap/selectors'
@@ -10,6 +10,7 @@ import { Field } from 'src/swap/types'
 import { useTokensWithTokenBalance } from 'src/tokens/hooks'
 import { TokenBalance } from 'src/tokens/slice'
 import { NetworkId } from 'src/transactions/types'
+import { getSupportedNetworkIds } from 'src/web3/utils'
 
 export default function useFilterChip(
   selectingField: Field | null,
