+++ src/transactions/saga.ts
@@ -5,7 +5,6 @@ import { EarnEvents, SwapEvents } from 'src/analytics/Events'
 import { trackPointsEvent } from 'src/points/slice'
 import { earnPositionsSelector } from 'src/positions/selectors'
 import { RootState } from 'src/redux/store'
-import { getSupportedNetworkIds } from 'src/statsig'
 import { tokensByIdSelector } from 'src/tokens/selectors'
 import { BaseToken } from 'src/tokens/slice'
 import { transactionFeedV2Api, TransactionFeedV2Response } from 'src/transactions/api'
@@ -25,6 +24,7 @@ import {
 import Logger from 'src/utils/Logger'
 import { publicClient } from 'src/viem'
 import networkConfig, { networkIdToNetwork } from 'src/web3/networkConfig'
+import { getSupportedNetworkIds } from 'src/web3/utils'
 import { call, delay, fork, put, select, spawn, takeEvery } from 'typed-redux-saga'
 import { Hash, TransactionReceipt } from 'viem'
